requirements.txt
# Core dependencies
# CRITICAL: notion-client 2.7.0 has breaking change - removes databases.query() method!
# Pin to 2.2.1 until upstream fix or workaround implemented
notion-client==2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Advanced fuzzy matching for deduplication
rapidfuzz>=3.0.0

# Audio processing
yt-dlp>=2023.12.30
librosa>=0.10.1
numpy>=1.24.0
soundfile>=0.12.1
mutagen>=1.47.0
send2trash>=1.8.2











































































































